Eric Sunshine made the following changes:
-*- Eliminated compilation warnings about "non-virtual destructor in class
    with virtual methods" from the following classes: csModelDataObject,
    csModelDataCamera, csModelDataLight, csModelDataMaterial, and
    csModelData.

-*- Eliminated several compilation warnings about implicit conversion from
    double to integer in csEngine::SetAmbientLight().


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7433 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csengine/engine.cpp
@@ -2387,9 +2387,9 @@ iCameraPosition* csEngine::GetCameraPosition (const char* name) const
 
 void csEngine::SetAmbientLight (const csColor &c)
 {
-  csLight::ambient_red = c.red * 255.0;
-  csLight::ambient_green = c.green * 255.0;
-  csLight::ambient_blue = c.blue * 255.0;
+  csLight::ambient_red = int(c.red * 255);
+  csLight::ambient_green = int(c.green * 255);
+  csLight::ambient_blue = int(c.blue * 255.0);
 }
 
 void csEngine::GetAmbientLight (csColor &c) const
